Add support for spill suppression to avg_chromakey.py.
svn path=/trunk/libavg/; revision=5178

+++ src/player/ChromaKeyFXNode.cpp
@@ -130,7 +130,7 @@ void ChromaKeyFXNode::setSpillThreshold(double spillThreshold)
     setFilterParams();
 }
 
-int ChromaKeyFXNode::getSpillThreshold() const
+double ChromaKeyFXNode::getSpillThreshold() const
 {
     return m_SpillThreshold;
 }
